Add docs, PostgreSQL migration, tests & server APIs
Major update preparing PPA for V2 and a PostgreSQL migration: add extensive docs and API documentation (docs/, server/docs/, server/swagger/), introduce migration scripts and index tuning (server/scripts/migration/), and add a large suite of PostgreSQL-related and backup tests (server/tests/, tests/postgresql-query-performance/). Update server to expose new API docs and routes (server/routes/, server/controllers/, server/services/, server/models/), adjust package files, and add helper scripts for generating swagger/docs. Polish developer experience: README rewritten for V2, expanded BMAD skill docs (bmad-brainstorming, bmad-party-mode), update .gitignore to hide DB/backups, add .mcp.json and .claude local settings, and add test orchestration scripts. These changes prepare the codebase for Supabase/Postgres, improve testing/performance tooling, and centralize API docs and examples.

+++ server/init-db.js
@@ -10,7 +10,7 @@ const CREATE_TABLES_SQL = `
       is_template BOOLEAN NOT NULL DEFAULT 0,
       project_type TEXT DEFAULT 'standard',
       final_total_cost REAL,
-      final_risk_score INTEGER,
+      final_risk_score REAL,
       final_workload_days REAL,
       assessment_details_json TEXT,
       tags_json TEXT,
